Add test vectors for the new history network (#56)
These "block data" test vectors just included the header, body
and receipts rlp encoded data.

The body and receipts as they are to be send over the Portal
history network. Header is included to be able to also verify
them in tests.

No content key is added as this is rather trivial and covered
by other test vectors.